chore(backport): backport all necessary changes to make the CI work on push (#31)
* ci: fix incorrect workflow name (#19)

The ci.yaml uses integration_test.yaml instead of integration_tests.yaml

* ci: add workflow_call to integration_test.yaml (#21)

This allows the workflow to be called from other workflows.

* ci: publish to 1.23 from the default branch (#26)

* ci: publish to 1.23 from the default branch

This commit forces the release workflow to publish to the 1.23/edge track from the default branch.

* chore: publish to latest/edge by default

* chore: add upstream-source to resources (#27)

This commit adds the upstream-source to the resources item in the metadata.yaml file.